+++ apps/quickstart/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @walkeros/quickstart
 
+## 4.6.1
+
+### Patch Changes
+
+- @walkeros/collector@4.6.1
+- @walkeros/core@4.6.1
+- @walkeros/web-core@4.6.1
+- @walkeros/web-destination-api@4.6.1
+- @walkeros/web-destination-gtag@4.6.1
+- @walkeros/web-source-browser@4.6.1
+
 ## 4.6.0
 
 ### Patch Changes