+++ packages/web/sources/demo/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @walkeros/source-demo
 
+## 4.6.1
+
+### Patch Changes
+
+- @walkeros/core@4.6.1
+
 ## 4.6.0
 
 ### Patch Changes